requirements.txt
streamlit>=1.24.0
google-cloud-aiplatform>=1.38.0
pillow>=9.0.0